Refactor: reorganized files
